Refactor:Remove ReadingList Labor Class, Replace with user Method
