--- app/labor/reading_list.rb
@@ -1,21 +1,0 @@
-class ReadingList
-  attr_accessor :user
-
-  def initialize(user)
-    @user = user
-  end
-
-  def cached_ids_of_articles
-    Rails.cache.fetch("reading_list_ids_of_articles_#{user.id}_#{user.public_reactions_count}") do
-      ids_of_articles
-    end
-  end
-
-  def ids_of_articles
-    Reaction.where(reaction_criteria).where.not(status: "archived").order(created_at: :desc).pluck(:reactable_id)
-  end
-
-  def reaction_criteria
-    { user_id: user.id, reactable_type: "Article", category: "readinglist" }
-  end
-end
